Remove text component finally
